refactor project roles part of create project
