+++ fe/src/features/projects/CreateProject.vue
@@ -1,7 +1,7 @@
 <template>
   <div style="align-items: center; justify-content: center">
     <q-form
-      class="createProjectForm"
+      class="createProjectForm q-py-lg"
       @submit="onSubmit"
     >
       <div>
@@ -10,7 +10,7 @@
             Create project
           </div>
           <div>
-            <div class="q-gutter-sm">
+            <div>
               <btn
                 elevated
                 color="primary"
@@ -40,321 +40,26 @@
 
           <ProjectCoaches />
 
-          <div
-            class="projectcol col-xs-12 col-sm-12 col-md-12 col-lg-5 col-xl-6"
-          >
-            <h4 class="projectsubtitle">
-              Project Roles
-            </h4>
-            <div class="row">
-              <btn
-                class="cornered"
-                color="primary"
-                icon="add"
-                label="Add role"
-                @click="new_role_prompt = true"
-                glow-color="#00F1AF"
-                shadow-strength=2
-              />
-              <q-space />
-              <q-input
-                v-model="filter_roles"
-                style="max-width: 190px"
-                outlined
-                dense
-                debounce="300"
-                color="green"
-                class="inputfield"
-                placeholder="Search"
-                @keydown.enter.prevent=""
-              >
-                <template #append>
-                  <q-icon
-                    v-if="filter_roles !== ''"
-                    name="close"
-                    class="cursor-pointer"
-                    @click="filter_roles = ''"
-                  />
-                  <q-icon
-                    v-if="filter_roles === ''"
-                    name="search"
-                  />
-                </template>
-              </q-input>
-            </div>
-            <q-table
-              class="table shadow-4"
-              :rows="skillStore.skills"
-              :columns="columns_roles"
-              :loading="skillStore.isLoadingSkills"
-              row-key="name"
-              :filter="filter_roles"
-            >
-              <template #body="props">
-                <q-tr
-                  :class="props.rowIndex % 2 === 1 ? 'bg-green-1' : ''"
-                  :props="props"
-                >
-                  <q-td
-                    key="role"
-                    :props="props"
-                  >
-                    {{ props.row.name }}
-                  </q-td>
-                  <q-td
-                    key="amount"
-                    :props="props"
-                  >
-                    {{ props.row.amount }}
-                    <q-popup-edit
-                      v-slot="scope"
-                      v-model.number="props.row.amount"
-                      buttons
-                      label-set="Save"
-                      label-cancel="Close"
-                      :validate="amountRangeValidation"
-                    >
-                      <q-input
-                        v-model.number="scope.value"
-                        type="number"
-                        hint="Enter a positive number."
-                        :error="errorRoleAmount"
-                        :error-message="errorMessageRoleAmount"
-                        dense
-                        autofocus
-                        borderless
-                        @keyup.enter="scope.set"
-                      />
-                    </q-popup-edit>
-                  </q-td>
-                  <q-td
-                    key="comment"
-                    :props="props"
-                  >
-                    <div>{{ props.row.comment }}</div>
-                    <q-popup-edit
-                      v-slot="scope"
-                      v-model="props.row.comment"
-                      buttons
-                    >
-                      <q-input
-                        v-model="scope.value"
-                        type="text"
-                        autogrow
-                        autofocus
-                        counter
-                        borderless
-                        @keyup.enter.stop
-                      />
-                    </q-popup-edit>
-                  </q-td>
-                  <q-td
-                    key="color"
-                    :props="props"
-                    auto-width
-                  >
-                    <div
-                      :style="`height: 25px; width:25px; border-radius: 50%;background: ${props.row.color}`"
-                    />
-                    <!-- TODO make this actually change in the database not locally-->
-                    <q-popup-edit
-                      v-slot="scope"
-                      v-model="props.row.color"
-                      buttons
-                    >
-                      <q-color
-                        v-model="scope.value"
-                        no-header
-                        no-footer
-                        class="color-picker"
-                        @keyup.enter.stop
-                      />
-                    </q-popup-edit>
-                  </q-td>
-                  <q-td
-                    key="remove"
-                    style="width: 10px"
-                  >
-                    <btn
-                      flat
-                      round
-                      style="color: #f14a3b"
-                      icon="mdi-trash-can-outline"
-                      glow-color="red-2"
-                      @click="delete_role = props.row"
-                    />
-                  </q-td>
-                </q-tr>
-              </template>
-            </q-table>
-          </div>
+          <ProjectRoles />
         </div>
       </div>
     </q-form>
   </div>
-
-  <q-dialog
-    v-model="new_role_prompt"
-    persistent
-  >
-    <q-card class="create-role-popup">
-      <q-card-section>
-        <div class="text-h6">
-          Create a new role
-        </div>
-      </q-card-section>
-
-      <q-card-section class="q-pt-none">
-        <q-input
-          v-model="new_role"
-          outlined
-          autofocus
-          class="inputfield"
-          label="Role name"
-          lazy-rules
-          :rules="[
-            (val) =>
-              (val && val.length > 0) || 'Enter the name of the new role.',
-          ]"
-        />
-      </q-card-section>
-      <q-card-section class="q-pt-none">
-        <!--        TODO REMOVE -->
-        <q-input
-          v-model="new_role_color"
-          outlined
-          label="text color"
-          class="inputfield"
-          type="url"
-        />
-        <!--  INFO if picker gives conversion issues use-->
-        <!--  INFO https://quasar.dev/quasar-utils/color-utils#color-conversion-->
-        <q-color
-          v-model="new_role_color"
-          no-header
-          no-footer
-          class="color-picker"
-        />
-      </q-card-section>
-      <q-card-actions
-        align="right"
-        class="text-primary"
-      >
-        <btn
-          v-close-popup
-          flat
-          label="Cancel"
-          glow-color="#C0FFF4"
-        />
-        <btn
-          flat
-          label="Add role"
-          @click="new_role_confirm"
-          glow-color="#C0FFF4"
-        />
-      </q-card-actions>
-    </q-card>
-  </q-dialog>
-
-  <q-dialog
-    :model-value="delete_role !== undefined"
-    @update:model-value="delete_role=undefined"
-    persistent
-  >
-    <q-card style="min-width: 350px">
-      <q-card-section horizontal>
-        <q-card-section class="col-3 flex flex-center">
-          <q-icon
-            name="warning"
-            class="text-red"
-            size="80px"
-          />
-        </q-card-section>
-        <q-card-section class="q-pt-xs">
-          <div class="text-h6 q-mt-sm q-mb-xs">
-            Are you sure you want to delete "{{ delete_role?.name }}"?
-          </div>
-          <div class="text text-grey">
-            This skill will be deleted immediately from all projects. You can't
-            undo this action.
-          </div>
-        </q-card-section>
-      </q-card-section>
-
-      <q-card-actions
-        align="right"
-        class="text-primary"
-      >
-        <btn
-          v-close-popup
-          flat
-          color="grey"
-          label="Cancel"
-        />
-        <btn
-          flat
-          color="red"
-          label="Delete"
-          @click="delete_role_confirm(delete_role!.id)"
-          glow-color="red-2"
-        />
-      </q-card-actions>
-    </q-card>
-  </q-dialog>
 </template>
 
 <script lang="ts">
 import router from '../../router'
-import { useQuasar } from 'quasar'
-import { ref, Ref } from 'vue'
+import { ref } from 'vue'
 import { defineComponent, onMounted } from '@vue/runtime-core'
 import { useSkillStore } from '../../stores/useSkillStore'
 import { useCoachStore } from '../../stores/useCoachStore'
-import { User } from '../../models/User'
-import { SkillInterface } from '../../models/Skill'
 import BasicInfo from "./components/BasicInfo.vue";
 import {useProjectStore} from "../../stores/useProjectStore";
 import ProjectCoaches from "./components/ProjectCoaches.vue";
-const columns_roles = [
-  {
-    name: 'role',
-    align: 'left' as const,
-    label: 'Project Role',
-    field: 'role',
-    sortable: true,
-  },
-  {
-    name: 'amount',
-    align: 'left' as const,
-    label: 'Amount',
-    field: 'amount',
-    sortable: true,
-  },
-  {
-    name: 'comment',
-    align: 'left' as const,
-    label: 'Comment',
-    field: 'comment',
-    sortable: true,
-  },
-  {
-    name: 'color',
-    align: 'center' as const,
-    label: 'Color',
-    field: 'color',
-    sortable: false,
-  },
-  {
-    name: 'action',
-    align: 'right' as const,
-    label: '',
-    field: '',
-    sortable: false,
-  },
-]
+import ProjectRoles from "./components/ProjectRoles.vue";
 
 export default defineComponent({
-  components: {ProjectCoaches, BasicInfo},
+  components: {ProjectCoaches, BasicInfo, ProjectRoles},
   setup() {
     const skillStore = useSkillStore()
     const coachStore = useCoachStore()
@@ -365,24 +70,6 @@ export default defineComponent({
       coachStore.loadUsers()
     })
 
-    const q = useQuasar()
-
-    // Role amount error handling
-    const errorRoleAmount = ref(false)
-    const errorMessageRoleAmount = ref('')
-
-    // Filters
-    const filter_roles = ref('')
-    const filter_coaches = ref('')
-
-    // variables for the new role dialog popup
-    const new_role_prompt = ref(false)
-    const new_role = ref('')
-    const new_role_color = ref('')
-
-    // variables for the delete role dialog popup
-    const delete_role: Ref<SkillInterface|undefined> = ref()
-
     const selected_coaches = ref([])
 
     return {
@@ -390,87 +77,7 @@ export default defineComponent({
       coachStore,
       projectStore,
 
-      filter_roles,
-      filter_coaches,
-
       selected_coaches,
-      columns_roles,
-      /*
-       * Role amount validation
-       */
-      errorRoleAmount,
-      errorMessageRoleAmount,
-      amountRangeValidation(val: number) {
-        if (val < 0) {
-          errorRoleAmount.value = true
-          errorMessageRoleAmount.value = 'The value must be positive!'
-          return false
-        }
-        errorRoleAmount.value = false
-        errorMessageRoleAmount.value = ''
-        return true
-      },
-      delete_role,
-
-      delete_role_confirm(id: number) {
-        skillStore.deleteSkill(id)
-        
-        delete_role.value = undefined
-
-        q.notify({
-          icon: 'done',
-          color: 'positive',
-          message: 'Successfully deleted!',
-        })
-      },
-
-      /*
-       * New Role
-       */
-      new_role_prompt,
-      new_role,
-      new_role_color,
-      new_role_confirm() {
-        // check if the new role value is valid
-        if (
-          new_role.value &&
-          new_role.value.length > 0 &&
-          new_role_color.value.length > 0
-        ) {
-          // when valid call the store object and add the skill
-          skillStore.addSkill(
-            new_role.value,
-            new_role_color.value,
-            // callback
-            (success: boolean) => {
-              if (success) {
-                q.notify({
-                  icon: 'done',
-                  color: 'positive',
-                  message: `Added new project role: ${new_role.value}.`,
-                  textColor: 'black',
-                })
-                new_role_prompt.value = false
-                new_role.value = ''
-                new_role_color.value = ''
-              } else {
-                q.notify({
-                  icon: 'close',
-                  color: 'negative',
-                  message: 'Failed to add role!',
-                })
-              }
-            }
-          )
-        } else {
-          q.notify({
-            icon: 'close',
-            color: 'negative',
-            message: 'Invalid name/color!',
-          })
-        }
-      },
-
     }
   },
   methods: {
@@ -518,9 +125,8 @@ thead {
     padding-right: 15px
 
 .projectsubtitle
+    margin-block-start: 0.5em !important
     font-weight: 300
-    margin-top: 10px
-    margin-bottom: 15px
     font-size: x-large
 
 .table
@@ -528,7 +134,6 @@ thead {
     margin-top: 10px
 
 .createProjectForm
-    margin-top: 25px
     margin-left: 10%
     margin-right: 10%
